+++ R/GramSchmidt.R
@@ -8,26 +8,60 @@
 #'
 #' @param X a matrix
 #' @param normalize logical; should the resulting columns be normalized to unit length?
+#' @param verbose logical; if \code{TRUE}, print intermediate steps
 #' @return A matrix of the same size as \code{X}, with orthogonal columns
 #' @author Phil Chalmers
 #' @export
 #' @examples
+#' (xx <- matrix(c( 1:3, 3:1, 1, 0, -2), 3, 3))
+#' crossprod(xx)
+#' (zz <- GramSchmidt(xx, normalize=FALSE))
+#' zapsmall(crossprod(zz))
+#' 
+#' # normalized
+#' (zz <- GramSchmidt(xx))
+#' zapsmall(crossprod(zz))
+#' 
+#' # print steps
+#' GramSchmidt(xx, verbose=TRUE)
+#' 
+#' # system with no unique solution
 #' (xx <- matrix(c( 1:3, 3:1, 1, 0, -1), 3, 3))
 #' crossprod(xx)
-#' (yy <- GramSchmidt(xx))
-#' zapsmall(crossprod(yy))
-
-GramSchmidt <- function(X, normalize=TRUE){
+#' (zz <- GramSchmidt(xx))
+#' zapsmall(crossprod(zz))
+#' 
+GramSchmidt <- function(X, normalize=TRUE, verbose=FALSE){
   B <- X
+  B[,2L:ncol(B)] <- 0
+  if (verbose){
+  	cat("\nInitial matrix:\n")
+  	printMatrix(X)
+  	cat("\nColumn 1: z1 = x1\n")
+  	printMatrix(B)
+  }
   for(i in 2:ncol(X)){
   	res <- X[,i]
   	for(j in 1:(i-1))
   		 res <- res - Proj(X[,i],B[,j])
   	B[,i] <- res
+  	if (verbose){
+  		prj_char <- character(i-1)
+  		for(j in 1:(i-1))
+  			prj_char[j] <- sprintf('Proj(x%i, z%i)', i, j)
+  		cat(sprintf("\nColumn %i: z%i = x%i - %s\n", 
+  					i, i, i, paste0(prj_char, collapse=' - ')))
+  		printMatrix(B)
+  	}
   }
   if (normalize) {
-	  norm <- diag(1/sqrt(colSums(B^2)))
+	  norm <- diag(1/len(B))
 	  B <- B %*% norm
+	  if (verbose){
+	  	cat("\nNormalized matrix: Z * inv(L) \n")
+	  	printMatrix(B)
+	  }
   }
+  if(verbose) return(invisible(B))
   B
 }